* fixed the last issue with the intel opencl compiler/runtime on linux (working via x11 forwarding on os x: https://dl.dropboxusercontent.com/u/1004620/oclr_intel_linux_x11fwd.png)  * print some more detailed error messages when setting a kernel argument failed

+++ lib/pipeline/rasterization_stage.cpp
@@ -74,7 +74,7 @@ void rasterization_stage::rasterize(draw_state& state,
 	ocl->set_kernel_argument(argc++, (unsigned int)(state.bin_count.x * state.bin_count.y));
 	ocl->set_kernel_argument(argc++, state.bin_offset);
 	ocl->set_kernel_argument(argc++, state.batch_count);
-	ocl->set_kernel_argument(argc++, intra_bin_groups);
+	ocl->set_kernel_argument(argc++, (unsigned int)intra_bin_groups);
 	ocl->set_kernel_argument(argc++, (underlying_type<PRIMITIVE_TYPE>::type)type);
 	ocl->set_kernel_argument(argc++, state.instance_primitive_count);
 	ocl->set_kernel_argument(argc++, state.instance_index_count);
